Add favicon
Fun fact: if a <link rel="icon"> is not found, browsers will automatically make a request to `./favicon.ico` or something like that. The 404 page for GitHub Pages is several kilobytes of HTML, so by including an icon we actually transfer less data. That doesn't actually matter, I just found it funny.